deeplabv3/train_plates/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.20.0
Pillow>=8.0.0
tqdm>=4.60.0
opencv-python>=4.5.0
pandas>=1.3.0
